move jsons

+++ bot_modules/adblock.js
@@ -4,17 +4,18 @@ exports.on = async function(event, i) {
     switch (event) {
         case "message":
             if (i.v.guilds_adblock.includes(i.m.guild.id)) {
-                if (i.m.content.includes("http://") ||
+                if ((i.m.content.includes("http://") ||
                     i.m.content.includes("https://") ||
                     i.m.content.includes("discord.gg") ||
-                    i.m.content.includes("discord.io")) {
+                    i.m.content.includes("discord.io")) &&
+                    i.m.author.id != bot.user.id) {
                     if (i.m.deletable) { i.m.delete(); } 
                     else { i.m.channel.send(`I am missing \`MANAGE_MESSAGES\` permissions for adblock;\nIf you did not intend to enable adblock, type \`${i.config.prefix}adblock\`.`); }
                 }
             }
             break
         case "save":
-            fs.writeFileSync("./guilds_adblock.json", JSON.stringify(i.v.guilds_adblock));
+            fs.writeFileSync("./../json/guilds_adblock.json", JSON.stringify(i.v.guilds_adblock));
             break
         case "guildDelete":
             for (var u = 0; u < imports.v.guilds_adblock.length; u++) {
